extract morph action and add children-only option to stream-element
